Moved commands Initalize client in rootCmd
