+++ cmd/add.go
@@ -1,18 +1,27 @@
-package commands
+package cmd
 
 import (
 	api "github.com/8Mobius8/go-habits/api"
 	"github.com/spf13/cobra"
-	"github.com/spf13/viper"
 )
+
+func init() {
+	rootCmd.AddCommand(addCmd)
+}
+
+var addCmd = &cobra.Command{
+	Use:     "add",
+	Short:   "Add a todo to Habitica",
+	Aliases: []string{"a", "a t"},
+	Run:     Add,
+}
 
 // Add or `go-habits add` command allows habiters to add
 // new tasks to their list of todos. When running this command
 // the format expected is like follows:
 // go-habits a {{ TaskTitle }}
 func Add(cmd *cobra.Command, args []string) {
-	client := api.NewHabiticaAPI(nil, viper.GetString("server"))
-	client.UpdateUserAuth(getAuthConfig())
+	client := habitsServer
 
 	t := api.Todo{}
 	t.Title = parseTodoTitle(args)
