+++ cmd/add_test.go
@@ -1,4 +1,4 @@
-package commands
+package cmd
 
 import (
 	"math/rand"
